Notes
clerical error on page 203 on line 4.  There is an additional ' ) ' at the end of [[prototype]]. Had to create many of my own examples as the author used the same code (eg: variables names )  for different examples hence would not have been able to run the script due to errors.   Provided output log for all examples without ( it seems a little random when he will console log some examples and others will just provide the output on the page )