Notes
On page 313 regarding the Handy Dandy Function it really should be insertAfter (h1Element,newElement),  even though technically the code will  work with the BodyElement, you will be adding a sibling to the BodyElement but throughout the last few examples we have been working with the children of the BodyElement hence it is much more clear and straightforward. It also makes finding the h1Element pointless if 'BodyElement' is the correct argument the author intended to use.