Notes
On page 313 regarding the Handy Dandy Function it really should be insertAfter (h1Element,newElement),  even though technically the code will  work with the BodyElement, you will be adding a sibling to the BodyElement but throughout the last few examples we have been working with the children of the BodyElement hence it will be  much more clear and straightforward working with the children elements.  The author really only needed to find the bodyElement or the h1Element depending which he was going to use when he invoked the function and not both.